refactor(iperf): decompose the iperf god-file by role (WS-D)
Split internal/diagnostics/iperf/iperf.go (815 LOC) into four role-based
files in the same package — pure move:

- iperf.go   shared types, Manager constructor, status getters
- binary.go  iperf3 binary discovery + version checks
- server.go  server lifecycle (start/stop)
- client.go  client test execution + result parsing

All four under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
